Extract Engine module and add plain HTTP /api/exec endpoint
Separate transport-agnostic V8 execution logic into engine/ module
so both MCP (ServerHandler) and plain HTTP (axum) can share it.
This enables k6 load testing without SSE stream complications.

- engine/mod.rs: Engine struct with run_js, list_sessions, snapshots
- engine/heap_storage.rs, session_log.rs: moved from mcp/
- mcp.rs: thin McpService adapter wrapping Engine
- api.rs: POST /api/exec with JSON request/response
- main.rs: unified wiring, no more stateless/stateful duplication
- Load test files updated to use /api/exec
