+++ server/src/engine/session_log.rs
