+++ server/src/lib.rs
@@ -1,2 +1,4 @@
 pub mod cluster;
+pub mod engine;
 pub mod mcp;
+pub mod api;
